Add country field on participant profile page
Part of issue #353.
The  country field was added because it is part of
the CLEF DB structure. It is shown in a new tab called
'Contact info' exclusively to the current participant, the organizers
and the admins on the participant profile page.

For the countries dropdown the gem country_select was added to the
.Gemfile

There was an existing location field in the database. This field was
removed in agreement with Sean.

In order to make this change, the gollowing was done:
 - Add country_cd field to db

 - Add select field to app/views/participants/edit.html.erb

 - Add country_cd field in accepted params for updates in
   app/controllers/participants_controller.rb

 - Remove existing location field from database, from
   app/views/participants/edit.html.erb and location_organization from
   app/views/participants/edit.html.erb

 - Add validation for country_code to either accept a blank value or a
   value contained in the list of country codes from countries gem in
   app/models/participant.rb

 - Add contact info tab which shows email and country of participant
   in app/views/participants/show.html.erb