Add address field on participant profile page
Part of issue #353.

The address field was added because it is part of
the CLEF DB structure. In order to show it, it was also added to
the 'Contact info' tab, exclusively viewable by the current
participant, the organizers and the admins on the participant profile
page.

Validation for address length (between 10 and 255 chars) was added.